Remove use of path.join for manifest.launch_path
I was having issues on Windows when creating a new FirefoxOS project. A brand new project, when added to the Firefox App Manager, gave an error: "Launch path has to be an absolute path starting with '/': '\index.html'". The issue is the use of path.join to generate manifest.launch_path. On Windows, this uses backslashes, which caused the issue. I have opted to simply use concatenation instead.

I have also fixed an issue where "cordova prepare firefoxos" would fail if there was no <content> tag in the config.xml. If it is not present, we default to index.html.