Fix `cordova platform add` not using searchpath from config.json
This was causing a problem when a platform was added and one of the already
installed plugins had a dependency specific to this new platform (this is
rather rare).

One example is org.chromium.socket depends on org.chromium.iossocketscommon
only for iOS but not for Android.

To reproduce using mobile-chrome-apps:

    cca create TstApp --link-to=spec --android
    cd TstApp
    cca prepare
    cca platform add ios