CB-9065 Allow removing plugins by short name.
Because the 'short name' (defined in plugin.xml) might go away if we ever switch to using package.json instead of plugin.xml, the logic used here is instead to allow ommitting 'cordova-plugin-' from the beginning of a plugin name when removing. Note that this will also work if a plugin known to the plugin mapper is installed using the old style name.